fixed wrong type of Delta
git-svn-id: svn+ssh://svn.yandex.ru/yt/trunk@1086 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ yt/ytlib/misc/metric.h
@@ -45,7 +45,7 @@ private:
     double SumSquares;
 
     //! The length of segment correpsonding to one bucket.
-    int Delta;
+    double Delta;
 
     //! Contains number of values from (-inf, #MinValue).
     int MinBucket;
